+++ app/data/get-organisation.js
@@ -1,0 +1,33 @@
+const Wreck = require('@hapi/wreck')
+const { serverConfig } = require('../config')
+const { AUTH_COOKIE_NAME } = require('../constants/cookies')
+
+const getOrganisation = async (request) => {
+  const query = `query {
+          organisation(organisationId: ${request.auth.credentials.organisationId}) {
+            sbi
+            name
+            mobile
+            email
+            address {
+              fullAddress
+            }
+            type
+            legalStatus
+          }
+        }`
+
+  const { payload } = await Wreck.post(serverConfig.dataHost, {
+    headers: {
+      crn: request.auth.credentials.crn,
+      Authorization: request.state[AUTH_COOKIE_NAME],
+      'Content-Type': 'application/json'
+    },
+    payload: JSON.stringify({ query }),
+    json: true
+  })
+
+  return payload.data.organisation
+}
+
+module.exports = getOrganisation
