+++ app/data/index.js
@@ -1,0 +1,9 @@
+const getNotification = require('./get-notification')
+const getNotifications = require('./get-notifications-by-sbi')
+const getOrganisation = require('./get-organisation')
+
+module.exports = {
+  getNotification,
+  getNotifications,
+  getOrganisation
+}
